requirements.txt
streamlit==1.47.1
streamlit-drawable-canvas==0.9.3
tensorflow==2.19.0
pillow==11.3.0
numpy==2.1.3
pandas==2.3.1
scikit-learn==1.7.1
matplotlib==3.10.3
